backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
passlib>=1.7.4
bcrypt>=4.0.0
argon2-cffi>=21.3.0
PyJWT>=2.8.0
python-multipart>=0.0.6
httpx>=0.26.0
